+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -9,29 +9,29 @@ public interface FilmService {
     /**
      * Проставляем лайк фильму от пользователя
      */
-    void addLike(int filmId, int userId);
+    void addLike(long filmId, long userId);
 
     /**
      * Удаляем лайк у фильма от пользователя
      */
-    void removeLike(int filmId, int userId);
+    void removeLike(long filmId, long userId);
 
     /**
      * Возвращаем отсортированный список фильмов по рейтингу
      * @param count число фильмов, которые надо вернуть
      * @return отсортированный список фильмов
      */
-    List<Film> returnPopularFilms(int count);
+    List<Film> returnPopularFilms(long count);
 
     Film addFilm(Film film);
 
     Film updateFilm(Film film);
 
-    void removeFilmById(int filmId);
+    void removeFilmById(long filmId);
 
     void removeAllFilms();
 
-    Film retrieveFilmById(int filmId);
+    Film retrieveFilmById(long filmId);
 
     List<Film> retrieveAllFilms();
 }
